libalpm: reimport expired keys
If the user does not update for a while some of the keys in the keyring
may expire. Pacman does not import new versions of these keys because
they are already in the keying. This leads to users needing to first
update their keyring to get the new keys.

So when keys are expired, prompt to reimport them, in hopes that there's
a newer version of the key on the keyserver.