Remove initialisation of unions
As of gcc-15, unions are no longer zero initialised when "{0}" is used.
To revert to the old behaviour, either assumes building for C23 or
adding a compiler option that may or may not work with other compilers.

Remove current "initialization" to make it clear that full initialization
is not occuring. All relevant fields are currently initialized before
use.

Signed-off-by: Allan McRae <allan@archlinux.org>